% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/growth_form.R
\name{growth_form}
\alias{growth_form}
\alias{growth_form_del}
\alias{growth_form_version_current}
\alias{growth_form_versions}
\title{Plant growthform lookup table}
\usage{
growth_form(version = NULL)

growth_form_versions(type = "local")

growth_form_version_current(type = "local")

growth_form_del(version)
}
\arguments{
\item{version}{Version number.  The default will load the most
  recent version on your computer or the most recent version known
  to the package if you have never downloaded the data before.
  With \code{growth_form_del}, specifying \code{version=NULL}
  will delete \emph{all} data sets.}

\item{type}{Type of version to return: options are "local"
(versions installed locally) or "github" (versions available on
github).  With any luck, "github" is a superset of "local".  For
\code{growth_form_version_current}, if "local" is given, but there
are no local versions, then we do check for the most recent
github version.}
}
\description{
Lookup table for plant growth form
}
\details{
These data are then currated--please flag any errors or potential additions on github
are at https://github.com/wcornwell/growthformdatabase
}
\examples{
#
# see the format of the resource
#
head(growth_form())
}

