% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{growth_form_lookup_table}
\alias{growth_form_lookup_table}
\title{Build a lookup table for a set of species, connecting the species names to growth form}
\usage{
growth_form_lookup_table(species_list, lookup_table = NULL)
}
\arguments{
\item{species_list}{Character vector of species bionomials  Genus and species may be seperated by  " "
or "_"}

\item{lookup_table}{Any growth form lookup table, but by default \code{\link{growth_form}}}
}
\description{
Build a lookup table for a set of species, connecting the species names to growth form
}
